Ability to hide tally
Also moved CSS to app.css
