+++ src/stores/gameSettings.js
@@ -6,6 +6,7 @@ export const useGameSettingsStore = defineStore('gameSettings', {
   state: () => ({
     removeFont: useLocalStorage('removeFont', false),
     disableSound: useLocalStorage('disableSound', false),
+    hideTally: useLocalStorage('hideTally', false),
 
     streamName: ''
   }),
